refactor(models): extract API schemas from routes/games.py into models/game.py
Move 10 Pydantic request/response models (CreateGameRequest,
SubmitTurnRequest, GameData, GameSummary, TurnAccepted, SaveResult,
UpdateGameRequest, GameEndedData, DeleteGameRequest, PaginationMeta)
and _ZERO_WIDTH_CHARS constant from routes/games.py into the models
layer. Update imports in routes and 2 test files.

Reduces routes/games.py from 2,191 to 2,104 lines. Models now live
alongside existing GameSession/GameState/GameStatus domain models.
