+++ src/tta/models/__init__.py
@@ -1,6 +1,20 @@
 """TTA domain models."""
 
-from tta.models.game import GameSession, GameState, GameStatus
+from tta.models.game import (
+    CreateGameRequest,
+    DeleteGameRequest,
+    GameData,
+    GameEndedData,
+    GameSession,
+    GameState,
+    GameStatus,
+    GameSummary,
+    PaginationMeta,
+    SaveResult,
+    SubmitTurnRequest,
+    TurnAccepted,
+    UpdateGameRequest,
+)
 from tta.models.player import Player, PlayerSession
 from tta.models.turn import (
     ParsedIntent,
@@ -12,15 +26,25 @@ from tta.models.turn import (
 )
 
 __all__ = [
+    "CreateGameRequest",
+    "DeleteGameRequest",
+    "GameData",
+    "GameEndedData",
     "GameSession",
     "GameState",
     "GameStatus",
+    "GameSummary",
+    "PaginationMeta",
     "ParsedIntent",
     "Player",
     "PlayerSession",
+    "SaveResult",
+    "SubmitTurnRequest",
     "TokenCount",
+    "TurnAccepted",
     "TurnRequest",
     "TurnResult",
     "TurnState",
     "TurnStatus",
+    "UpdateGameRequest",
 ]
